// packages/frontend/src/server/features/layer2s/liveness/types.ts
import type { TableReadyValue } from '@l2beat/config'
import type { TrackedTxsConfigSubtype, UnixTime } from '@l2beat/shared-pure'

export type LivenessTimeRange = '30d' | '90d' | 'max'
export type LivenessAnomalyStatus = 'ongoing' | 'recovered'

export interface LivenessDataPoint {
  averageInSeconds: number
  minimumInSeconds: number
  maximumInSeconds: number
}

export interface LivenessAnomaly {
  start: UnixTime
  end: UnixTime | undefined
  status: LivenessAnomalyStatus
  durationInSeconds: number
  subtype: TrackedTxsConfigSubtype
  avgInterval: number
  isApproved: boolean
  failureMechanism: TableReadyValue | undefined
}

export interface LivenessDetails {
  '30d'?: LivenessDataPoint | null
  '90d'?: LivenessDataPoint | null
  max?: LivenessDataPoint | null
  syncedUntil: number
}

export interface LivenessProject {
  batchSubmissions?: LivenessDetails
  stateUpdates?: LivenessDetails
  proofSubmissions?: LivenessDetails
  anomalies: LivenessAnomaly[]
}

export type LivenessResponse = Record<string, LivenessProject>